requirements.txt
# Core Python packages for RAG and embeddings
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# PostgreSQL and database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=7.0.0

# Vector operations and embeddings
pgvector>=0.2.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# LangChain for RAG applications
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.0
tiktoken>=0.5.0

# Data visualization
plotly>=5.15.0
bokeh>=3.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Optional: For advanced ML tasks
# tensorflow>=2.13.0
# pytorch-lightning>=2.0.0
